frontend handle interaction between searchbox and filters
